@@ -59,6 +59,7 @@
     "watch": "pkg-utils watch --tsconfig tsconfig.lib.json"
   },
   "engines": {
-    "node": ">=14"
-  }
+    "node": ">=14.18.0"
+  },
+  "gitHead": "a7b7dad6bdfde05883a61f78e998b2571dfbe0e2"
 }
