@@ -1,10 +1,12 @@
-const uniq = require('lodash/uniq')
-const transformPkgs = require('./utils/transformPkgs')
+import uniq from 'lodash/uniq'
+import {PackageManifest} from './types'
+
+import transformPkgs from './utils/transformPkgs'
 
 const COMMON_KEYWORDS = ['sanity', 'cms', 'headless', 'realtime', 'content']
-const supportedNodeVersionRange = '>=14.0.0'
+const supportedNodeVersionRange = '>=14.18.0'
 
-transformPkgs((pkgManifest) => {
+transformPkgs((pkgManifest: PackageManifest, {relativeDir}) => {
   const name = pkgManifest.name.split('/').slice(-1)[0]
 
   const engines = pkgManifest.engines
@@ -12,20 +14,24 @@ transformPkgs((pkgManifest) => {
     engines.node = supportedNodeVersionRange
   }
 
-  return {
-    ...pkgManifest,
-    engines,
-    author: 'Sanity.io <hello@sanity.io>',
+  const publishedFields = {
     bugs: {
       url: 'https://github.com/sanity-io/sanity/issues',
     },
     keywords: uniq(COMMON_KEYWORDS.concat(name).concat(pkgManifest.keywords || [])),
     homepage: 'https://www.sanity.io/',
-    license: 'MIT',
     repository: {
       type: 'git',
       url: 'git+https://github.com/sanity-io/sanity.git',
       directory: `packages/${pkgManifest.name}`,
     },
+  }
+
+  return {
+    ...pkgManifest,
+    engines,
+    author: 'Sanity.io <hello@sanity.io>',
+    license: 'MIT',
+    ...(pkgManifest.private ? {} : publishedFields),
   }
 })
