@@ -1,9 +1,10 @@
 /**
  * Copies the ./LICENSE-file to each package folder (warning - will overwrite every license file in all monorepo packages packages)
  */
-const {promises: fs} = require('fs')
-const path = require('path')
-const readPackages = require('./utils/readPackages')
+import {promises as fs} from 'fs'
+
+import path from 'path'
+import readPackages from './utils/readPackages'
 
 const readLicense = fs.readFile(path.join(process.cwd(), 'LICENSE'), 'utf-8')
 
@@ -13,7 +14,7 @@ Promise.all(
     const targetPath = path.join(pkg.dirname, 'LICENSE')
     await fs.writeFile(targetPath, licenseTxt)
     return targetPath
-  }),
+  })
 ).then((targetPaths) => {
   console.log('Copied the ./LICENSE file to %s to packages.', targetPaths.length)
 })
