@@ -1,18 +1,20 @@
 /* eslint-disable no-sync */
-const fs = require('fs')
-const path = require('path')
-const {getManifestPaths} = require('./getPackagePaths')
+import fs from 'fs'
+import path from 'path'
+import {Package} from '../types'
+import {getManifestPaths} from './getPackagePaths'
 
 const rootPath = path.join(__dirname, '..', '..')
 
-module.exports = function readPackages() {
+export default function readPackages(): Package[] {
   return getManifestPaths().map((file) => {
     const filePath = path.join(rootPath, file)
     const dirname = path.join(rootPath, path.dirname(file))
     return {
       path: filePath,
       dirname: dirname,
-      content: JSON.parse(fs.readFileSync(filePath, 'utf8')),
+      relativeDir: path.relative(rootPath, dirname),
+      manifest: JSON.parse(fs.readFileSync(filePath, 'utf8')),
     }
   })
 }
